+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.2.0 - 2018-10-27
+
+### Added
+
+- `NordeaEredFormatter` for formatting numbers as done in Nordea eRedovisning
+
 ## 2.1.0 - 2018-08-15
 
 ### Added